+++ components/Text/Text.tsx
@@ -1,5 +1,5 @@
 import { createText } from "@shopify/restyle";
-import { Theme } from "../themes/default";
+import { Theme } from "../../themes/default";
 
 const Text = createText<Theme>();
